refactor: split base_type_pending_sdk.rs into sub-modules
Split the 427-line monolithic file into base_type_pending_sdk/ directory
with sub-modules following the identity/ pattern:
- adapter.rs (81 lines): PendingSdkAdapter struct + BaseTypeFactory impl
- session.rs (65 lines): PendingSdkSession struct + BaseTypeSession impl
- tests.rs (288 lines): all 23 unit tests
- mod.rs (13 lines): re-exports PendingSdkAdapter

All files are ≤288 lines (well under the 400-line target).
All 23 tests pass. No clippy regressions.
